requirements.txt
grpcio==1.51.3
grpcio-tools==1.51.3
protobuf==4.22.1
python-dotenv==1.0.0
